+++ src/main/java/io/acari/DDLC/hax/IconHacker.kt
@@ -1,0 +1,6 @@
+package io.acari.DDLC.hax
+
+interface IconHacker {
+  fun handleMaterialThemeActivated(materialThemeActive: Boolean)
+  fun extractPatchersOnActivation()
+}
